# api/openapi.yaml
openapi: 3.1.0
info:
  title: skali API
  description: |
    Client-facing API of skali. Authentication is email+password with
    optional TOTP two-factor; sessions are opaque bearer tokens. The same
    API serves every client: the web BFF, the `skali` CLI, and future
    native apps.

    There is no registration endpoint: users are created by an instance
    admin (`/v1/users`) or by the server operator (`skalid user create`).
    Every user carries an instance-wide role, `admin` (everything,
    including users, nodes, and instance settings) or `member` (nothing
    until granted). Product access is per project and per environment on
    one ladder of roles, `none < read < deploy < maintain < admin`: a
    project membership is the member's default role on every environment
    of the project, a per-environment cell overrides it for one user, and
    an environment ceiling (`max_role`) caps what members inherit.
    Non-members do not see a project at all (`404`); a locked environment
    (`none`) is listed by name but its contents answer `403 forbidden`
    naming the required role. Project and environment payloads carry the
    caller's `access` so clients can disable and explain without guessing.
    All error responses share the `Error` envelope; clients should branch
    on `error.code`, not on messages.

    Security-critical operations sit behind sudo mode: they require the
    session to have proven the user's identity recently (within the
...

...

...
  version: 0.1.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/users:
    get:
      operationId: listUsers
      summary: List all users
      description: Admin only. Every user of the instance, oldest first.
      tags: [users]
      responses:
        "200":
          description: All users.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createUser
      summary: Create a user
      description: |
        Admin only. Provisions a user with an initial password; there is no
        self-service signup or invitation email by design. `role` defaults
        to `member`.
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                name:
                  type: string
                password:
                  type: string
                  minLength: 8
                role: { $ref: "#/components/schemas/Role" }
                create_projects:
                  type: boolean
                  default: false
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/users/{id}:
    patch:
      operationId: updateUser
      summary: Update a user's name, role, or create_projects
      description: |
        Admin only. Changing your own role is refused (`409`), as is
        demoting the last admin.
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: At least one property must be present.
              properties:
                name:
                  type: string
                role: { $ref: "#/components/schemas/Role" }
                create_projects:
                  type: boolean
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      operationId: deleteUser
      summary: Delete a user
      description: |
        Admin only. Deletes the user with their sessions and 2FA enrollment.
        Deleting yourself or the last admin is refused (`409`).
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/users/{id}/password:
    post:
      operationId: resetUserPassword
      summary: Reset a user's password
      description: |
        Admin only. Sets a new password without the old one — the recovery
        path, since there is no email-based reset. Revokes every session of
        the user.
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                new_password:
                  type: string
                  minLength: 8
      responses:
        "204":
          description: Password reset; all of the user's sessions are revoked.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token from login / 2FA verify.
  responses:
    BadRequest:
      description: Malformed request body or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid token, wrong credentials, or wrong code.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The caller is authenticated but lacks the role the operation needs
        (`error.code = forbidden`; the message names it, for example
        "maintain on environment production required"), or, on
        sudo-mode-gated operations, recent authentication is required
        (`error.code = reauth_required`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        The resource does not exist, or the caller is not a member of the
        project it belongs to (the two are indistinguishable by design).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The resource is in a state that does not allow the operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Role:
      type: string
      description: |
        Instance-wide role. Admins may do everything and manage users,
        nodes, and instance settings; members hold nothing until granted
        project membership.
      enum: [admin, member]
...
    User:
      type: object
      required: [id, email, name, role, create_projects, two_factor_enabled, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role: { $ref: "#/components/schemas/Role" }
        create_projects:
          type: boolean
          description: Whether a member may create projects (and so become their admin). Instance admins always may.
        two_factor_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - invalid_credentials
            - invalid_token
            - invalid_code
            - forbidden
            - reauth_required
            - not_found
            - conflict
            - version_conflict
            - invalid_manifest
            - rate_limited
            - node_unreachable
            - no_ready_pod
            - registry_disabled
            - registry_unavailable
            - internal
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
...